Handle update conflict in agentdeploy controller
We see these periodic error messages logged:

base_controller.go:159] "addon-deploy-controller" controller failed to sync
"prod-on-prem/submariner", err: Operation cannot be fulfilled on
managedclusteraddons.addon.open-cluster-management.io "submariner": the
object has been modified; please apply your changes to the latest version and
try again

Assuming no other controller is making frequent updates to the
ManagedClusterAddon, the most likely reason for the update conflicts is
due to caching. The controller obtains the ManagedClusterAddon from the
informer's cache so it may not get the most up-to-date version of the
resource which can cause the transient conflict errors when updating the
resource. This can happen if the resource is immediately re-queued after
an update and the updated version hasn't reached the cache yet when the
controller's sync function is invoked again.

The log messages are benign but they are disconcerting and can fill up
the log with noise. To avoid the logging, on conflict error, simply
re-queue the resource instead of returning the error.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>