Continue updating deployment manifests when uninstalling (#226)
This allows an addon to update its deployment manifests while the
predelete hook is running. This is useful when telling the addon to
enter uninstall mode.

Additionally, if there is a bug in the uninstall and it keeps the
uninstall from succeeding, we need to allow container image updates to move
forward in the event the update fixes the problem.

Relates:
https://issues.redhat.com/browse/ACM-8947

Signed-off-by: mprahl <mprahl@users.noreply.github.com>